fix: 收口网络栈状态对象
