+++ os/src/net/socket.rs
@@ -3,10 +3,8 @@
 use crate::sync::SpinLock;
 use crate::vfs::{File, FsError, InodeMetadata};
 use alloc::collections::VecDeque;
-use alloc::vec;
 use lazy_static::lazy_static;
-use smoltcp::iface::{Interface, SocketHandle as SmoltcpHandle, SocketSet};
-use smoltcp::socket::{tcp, udp};
+use smoltcp::iface::SocketHandle as SmoltcpHandle;
 use smoltcp::wire::{IpAddress, IpEndpoint, Ipv4Address};
 
 #[derive(Clone, Copy, Debug)]
@@ -18,127 +16,27 @@ pub enum SocketHandle {
 use alloc::collections::BTreeMap;
 use alloc::sync::Arc;
 
-pub struct NetIfaceWrapper {
-    device: SpinLock<crate::net::stack::NetDeviceAdapter>,
-    interface: SpinLock<Interface>,
-}
-
-impl NetIfaceWrapper {
-    pub fn poll(&self, sockets: &SpinLock<SocketSet<'static>>) -> bool {
-        let timestamp =
-            smoltcp::time::Instant::from_millis(crate::arch::timer::get_time_ms() as i64);
-        let mut dev = self.device.lock();
-
-        // 检查队列长度
-        let queue_len = dev.loopback_queue_len();
-        if queue_len > 0 {
-            crate::pr_debug!("poll: loopback queue has {} packets", queue_len);
-        }
-
-        let mut iface = self.interface.lock();
-        let mut sockets = sockets.lock();
-
-        crate::pr_debug!("poll: before iface.poll");
-        let result = iface.poll(timestamp, &mut *dev, &mut *sockets);
-        crate::pr_debug!("poll: result={:?}", result);
-
-        // NOTE: For loopback traffic, frames produced by Tx are enqueued into `loopback_queue`
-        // during this poll, and therefore won't be received until a subsequent poll. Do a small,
-        // bounded extra poll to consume newly enqueued frames, otherwise UDP workloads like
-        // iperf3 can appear to "stall" (server sees only the first datagram).
-        if dev.loopback_queue_len() > 0 {
-            const MAX_EXTRA_POLLS: usize = 2;
-            for _ in 0..MAX_EXTRA_POLLS {
-                if dev.loopback_queue_len() == 0 {
-                    break;
-                }
-                let _ = iface.poll(timestamp, &mut *dev, &mut *sockets);
-            }
-        }
-
-        // Drain UDP datagrams from shared per-port sockets and deliver them to per-fd queues.
-        //
-        // This must happen as part of the global poll path; otherwise, programs that wait in
-        // select()/poll() (e.g. iperf3 UDP server) never observe readability because our
-        // SocketFile::readable() checks the per-fd queue, not the smoltcp socket buffer.
-        //
-        // Lock order: SocketSet -> UDP_PORTS (see udp_attach_fd_to_port).
-        let delivered_udp = udp_dispatch_drain_locked(&mut *sockets);
-
-        // Reap TCP sockets that have finished a graceful close.
-        let mut pending = PENDING_TCP_CLOSE.lock();
-        pending.retain(|h| {
-            let state = sockets.get::<tcp::Socket>(*h).state();
-            if matches!(state, tcp::State::Closed | tcp::State::TimeWait) {
-                crate::pr_debug!(
-                    "[Socket] reap: removing closed tcp handle={:?}, state={:?}",
-                    h,
-                    state
-                );
-                sockets.remove(*h);
-                false
-            } else {
-                true
-            }
-        });
-
-        let changed = result != smoltcp::iface::PollResult::None || delivered_udp;
-        if changed {
-            crate::kernel::syscall::io::wake_poll_waiters();
-        }
-        changed
-    }
-
-    pub fn loopback_queue_len(&self) -> usize {
-        self.device.lock().loopback_queue_len()
-    }
-
-    pub fn with_context<F, R>(&self, f: F) -> R
-    where
-        F: FnOnce(&mut smoltcp::iface::Context) -> R,
-    {
-        let mut iface = self.interface.lock();
-        f(iface.context())
-    }
-}
-
 lazy_static! {
-    pub static ref FD_SOCKET_MAP: SpinLock<BTreeMap<(usize, usize), SocketHandle>> = SpinLock::new(BTreeMap::new());
-    // UDP port dispatcher: one smoltcp UDP socket per local port, plus multiple "logical" sockets
-    // (per fd) that receive datagrams based on their connected remote endpoint.
-    static ref UDP_PORTS: SpinLock<BTreeMap<u16, UdpPortEntry>> = SpinLock::new(BTreeMap::new());
-    // TCP sockets that initiated a graceful close on Drop, and should be removed
-    // from SocketSet once the close handshake completes.
-    //
-    // Lock order invariant: SocketSet -> PENDING_TCP_CLOSE (matches Drop path).
-    static ref PENDING_TCP_CLOSE: SpinLock<alloc::vec::Vec<SmoltcpHandle>> =
-        SpinLock::new(alloc::vec::Vec::new());
+    pub static ref FD_SOCKET_MAP: SpinLock<BTreeMap<(usize, usize), SocketHandle>> =
+        SpinLock::new(BTreeMap::new());
 }
-
-use crate::net::stack::{NET_IFACE, SOCKET_SET};
 
 use crate::uapi::fcntl::OpenFlags;
 use crate::uapi::socket::SocketOptions;
 
 const UDP_RXQ_CAP: usize = 64;
-const UDP_DGRAM_MAX: usize = 2048;
+pub(crate) const UDP_DGRAM_MAX: usize = 2048;
 
 #[derive(Debug)]
-struct UdpDatagram {
-    src: IpEndpoint,
-    len: usize,
-    data: [u8; UDP_DGRAM_MAX],
-}
-
-#[derive(Debug)]
-struct UdpPortEntry {
-    handle: SmoltcpHandle,
-    sockets: alloc::vec::Vec<alloc::sync::Weak<dyn crate::vfs::File>>,
+pub(crate) struct UdpDatagram {
+    pub(crate) src: IpEndpoint,
+    pub(crate) len: usize,
+    pub(crate) data: [u8; UDP_DGRAM_MAX],
 }
 
 pub struct SocketFile {
-    handle: SpinLock<Option<SocketHandle>>,
-    listen_sockets: SpinLock<alloc::vec::Vec<SocketHandle>>,
+    pub(crate) handle: SpinLock<Option<SocketHandle>>,
+    pub(crate) listen_sockets: SpinLock<alloc::vec::Vec<SocketHandle>>,
     listen_backlog: SpinLock<usize>,
     local_endpoint: SpinLock<Option<IpEndpoint>>,
     remote_endpoint: SpinLock<Option<IpEndpoint>>,
@@ -147,7 +45,7 @@ pub struct SocketFile {
     shutdown_wr: SpinLock<bool>,
     flags: SpinLock<OpenFlags>,
     options: SpinLock<SocketOptions>,
-    is_listener: SpinLock<bool>,
+    pub(crate) is_listener: SpinLock<bool>,
 }
 
 impl SocketFile {
@@ -232,9 +130,7 @@ impl SocketFile {
     /// This is used to provide a minimal accept/backlog behavior on top of smoltcp's
     /// single-socket listen model.
     pub fn take_established_from_listen_queue(&self) -> Option<SocketHandle> {
-        crate::net::stack::network_stack()
-            .lock()
-            .take_established_from_listen_queue(self)
+        crate::net::stack::network_stack().take_established_from_listen_queue(self)
     }
 
     pub fn set_handle(&self, new_handle: SocketHandle) {
@@ -257,11 +153,11 @@ impl SocketFile {
         *self.remote_endpoint.lock()
     }
 
-    fn udp_queue_len(&self) -> usize {
+    pub(crate) fn udp_queue_len(&self) -> usize {
         self.udp_rx_queue.lock().len()
     }
 
-    fn udp_push(&self, d: UdpDatagram) -> bool {
+    pub(crate) fn udp_push(&self, d: UdpDatagram) -> bool {
         let mut q = self.udp_rx_queue.lock();
         if q.len() == q.capacity() {
             return false;
@@ -270,7 +166,7 @@ impl SocketFile {
         true
     }
 
-    fn udp_pop(&self) -> Option<UdpDatagram> {
+    pub(crate) fn udp_pop(&self) -> Option<UdpDatagram> {
         self.udp_rx_queue.lock().pop_front()
     }
 
@@ -291,17 +187,13 @@ impl SocketFile {
     }
 
     pub fn is_closed(&self) -> bool {
-        crate::net::stack::network_stack()
-            .lock()
-            .socket_is_closed(self)
+        crate::net::stack::network_stack().socket_is_closed(self)
     }
 }
 
 impl Drop for SocketFile {
     fn drop(&mut self) {
-        crate::net::stack::network_stack()
-            .lock()
-            .drop_socket_file(self);
+        crate::net::stack::network_stack().drop_socket_file(self);
     }
 }
 
@@ -401,33 +293,23 @@ pub fn socket_sendto(
     buf: &[u8],
     endpoint: IpEndpoint,
 ) -> Result<usize, FsError> {
-    crate::net::stack::network_stack()
-        .lock()
-        .socket_sendto(handle, buf, endpoint)
+    crate::net::stack::network_stack().socket_sendto(handle, buf, endpoint)
 }
 
 impl File for SocketFile {
     fn readable(&self) -> bool {
-        crate::net::stack::network_stack()
-            .lock()
-            .socket_readable(self)
+        crate::net::stack::network_stack().socket_readable(self)
     }
     fn writable(&self) -> bool {
-        crate::net::stack::network_stack()
-            .lock()
-            .socket_writable(self)
+        crate::net::stack::network_stack().socket_writable(self)
     }
 
     fn read(&self, buf: &mut [u8]) -> Result<usize, FsError> {
-        crate::net::stack::network_stack()
-            .lock()
-            .socket_read(self, buf)
+        crate::net::stack::network_stack().socket_read(self, buf)
     }
 
     fn write(&self, buf: &[u8]) -> Result<usize, FsError> {
-        crate::net::stack::network_stack()
-            .lock()
-            .socket_write(self, buf)
+        crate::net::stack::network_stack().socket_write(self, buf)
     }
 
     fn metadata(&self) -> Result<InodeMetadata, FsError> {
@@ -448,401 +330,8 @@ impl File for SocketFile {
     }
 
     fn recvfrom(&self, buf: &mut [u8]) -> Result<(usize, Option<alloc::vec::Vec<u8>>), FsError> {
-        crate::net::stack::network_stack()
-            .lock()
-            .socket_recvfrom(self, buf)
+        crate::net::stack::network_stack().socket_recvfrom(self, buf)
     }
-}
-
-pub(crate) fn take_established_from_listen_queue_impl(file: &SocketFile) -> Option<SocketHandle> {
-    let sockets = SOCKET_SET.lock();
-    let mut q = file.listen_sockets.lock();
-    let mut i = 0;
-    while i < q.len() {
-        match q[i] {
-            SocketHandle::Tcp(h) => {
-                let s = sockets.get::<tcp::Socket>(h);
-                match s.state() {
-                    tcp::State::Established | tcp::State::CloseWait => {
-                        return Some(q.remove(i));
-                    }
-                    tcp::State::Closed => {
-                        q.remove(i);
-                        continue;
-                    }
-                    _ => {}
-                }
-            }
-            SocketHandle::Udp(_) => {
-                q.remove(i);
-                continue;
-            }
-        }
-        i += 1;
-    }
-    None
-}
-
-pub(crate) fn socket_is_closed_impl(file: &SocketFile) -> bool {
-    let sockets = SOCKET_SET.lock();
-    match file.handle.lock().as_ref() {
-        Some(SocketHandle::Tcp(h)) => {
-            let socket = sockets.get::<tcp::Socket>(*h);
-            socket.state() == tcp::State::Closed
-        }
-        _ => false,
-    }
-}
-
-pub(crate) fn drop_socket_file_impl(file: &SocketFile) {
-    let mut sockets = SOCKET_SET.lock();
-    if let Some(handle) = *file.handle.lock() {
-        match handle {
-            SocketHandle::Tcp(h) => {
-                let socket = sockets.get_mut::<tcp::Socket>(h);
-                let state = socket.state();
-                crate::pr_debug!("[Socket] Drop: handle={:?}, state={:?}", h, state);
-                // Active connections are closed gracefully and removed by the poll reaper.
-                match state {
-                    tcp::State::Closed | tcp::State::TimeWait | tcp::State::Listen => {
-                        sockets.remove(h);
-                    }
-                    _ => {
-                        crate::pr_debug!("[Socket] Drop: closing socket handle={:?}", h);
-                        socket.close();
-                        PENDING_TCP_CLOSE.lock().push(h);
-                    }
-                }
-            }
-            SocketHandle::Udp(h) => {
-                // UDP sockets may be shared by the per-port dispatcher.
-                let ports = UDP_PORTS.lock();
-                let is_shared = ports.values().any(|e| e.handle == h);
-                drop(ports);
-                if !is_shared {
-                    sockets.remove(h);
-                }
-            }
-        }
-    }
-    for handle in file.listen_sockets.lock().iter() {
-        match handle {
-            SocketHandle::Tcp(h) => {
-                sockets.remove(*h);
-            }
-            SocketHandle::Udp(h) => {
-                sockets.remove(*h);
-            }
-        }
-    }
-}
-
-pub(crate) fn socket_sendto_impl(
-    handle: SocketHandle,
-    buf: &[u8],
-    endpoint: IpEndpoint,
-) -> Result<usize, FsError> {
-    let result = {
-        let mut sockets = SOCKET_SET.lock();
-        match handle {
-            SocketHandle::Tcp(_) => Err(FsError::NotSupported),
-            SocketHandle::Udp(h) => {
-                let socket = sockets.get_mut::<udp::Socket>(h);
-                socket
-                    .send_slice(buf, endpoint)
-                    .map_err(|_| FsError::WouldBlock)?;
-                Ok(buf.len())
-            }
-        }
-    };
-    if result.is_ok() {
-        poll_network_interfaces_impl();
-        crate::kernel::syscall::io::wake_poll_waiters();
-    }
-    result
-}
-
-pub(crate) fn socket_readable_impl(file: &SocketFile) -> bool {
-    // Listener socket: only readable when a connection is ready to accept.
-    if *file.is_listener.lock() {
-        let sockets = SOCKET_SET.lock();
-
-        for handle in file.listen_sockets.lock().iter() {
-            if let SocketHandle::Tcp(h) = handle {
-                let s = sockets.get::<tcp::Socket>(*h);
-                if matches!(s.state(), tcp::State::Established | tcp::State::CloseWait) {
-                    return true;
-                }
-            }
-        }
-
-        if let Some(SocketHandle::Tcp(h)) = *file.handle.lock() {
-            let s = sockets.get::<tcp::Socket>(h);
-            return matches!(s.state(), tcp::State::Established | tcp::State::CloseWait);
-        }
-        return false;
-    }
-
-    let sockets = SOCKET_SET.lock();
-    match file.handle.lock().as_ref() {
-        Some(SocketHandle::Tcp(h)) => {
-            let socket = sockets.get::<tcp::Socket>(*h);
-            let can_recv = socket.can_recv();
-            let state = socket.state();
-            crate::pr_debug!(
-                "[Socket] readable: handle={:?}, state={:?}, can_recv={}",
-                h,
-                state,
-                can_recv
-            );
-            can_recv || matches!(state, tcp::State::Closed | tcp::State::CloseWait)
-        }
-        Some(SocketHandle::Udp(_)) => {
-            drop(sockets);
-            file.udp_queue_len() > 0
-        }
-        None => false,
-    }
-}
-
-pub(crate) fn socket_writable_impl(file: &SocketFile) -> bool {
-    let sockets = SOCKET_SET.lock();
-    match file.handle.lock().as_ref() {
-        Some(SocketHandle::Tcp(h)) => {
-            let socket = sockets.get::<tcp::Socket>(*h);
-            let can_send = socket.can_send();
-            let state = socket.state();
-            crate::pr_debug!(
-                "[Socket] writable: handle={:?}, state={:?}, can_send={}",
-                h,
-                state,
-                can_send
-            );
-            can_send
-        }
-        Some(SocketHandle::Udp(h)) => {
-            let socket = sockets.get::<udp::Socket>(*h);
-            socket.can_send()
-        }
-        None => false,
-    }
-}
-
-pub(crate) fn socket_read_impl(file: &SocketFile, buf: &mut [u8]) -> Result<usize, FsError> {
-    if file.is_shutdown_read() {
-        return Ok(0);
-    }
-
-    let mut sockets = SOCKET_SET.lock();
-    let result = match file.handle.lock().as_ref() {
-        Some(SocketHandle::Tcp(h)) => {
-            let socket = sockets.get_mut::<tcp::Socket>(*h);
-            let state = socket.state();
-            let recv_queue = socket.recv_queue();
-            crate::pr_debug!(
-                "[Socket] read: handle={:?}, state={:?}, recv_queue={}, buf.len()={}",
-                h,
-                state,
-                recv_queue,
-                buf.len()
-            );
-
-            if socket.state() == tcp::State::Closed {
-                return Ok(0);
-            }
-
-            if state == tcp::State::CloseWait && recv_queue == 0 {
-                return Ok(0);
-            }
-
-            let result = socket.recv_slice(buf).map_err(|_| FsError::WouldBlock);
-
-            if let Ok(0) = result {
-                if state == tcp::State::CloseWait {
-                    crate::pr_debug!(
-                        "[Socket] read: recv_slice returned 0 and state=CloseWait, returning EOF"
-                    );
-                    Ok(0)
-                } else if socket.may_recv() {
-                    crate::pr_debug!(
-                        "[Socket] read: recv_slice returned 0 but may_recv=true, returning EAGAIN"
-                    );
-                    Err(FsError::WouldBlock)
-                } else {
-                    crate::pr_debug!(
-                        "[Socket] read: recv_slice returned 0 and may_recv=false, returning EOF"
-                    );
-                    Ok(0)
-                }
-            } else {
-                if let Ok(n) = result {
-                    crate::pr_debug!("[Socket] read: received {} bytes", n);
-                }
-                result
-            }
-        }
-        Some(SocketHandle::Udp(_)) => {
-            drop(sockets);
-            let Some(d) = file.udp_pop() else {
-                return Err(FsError::WouldBlock);
-            };
-            let n = core::cmp::min(buf.len(), d.len);
-            buf[..n].copy_from_slice(&d.data[..n]);
-            Ok(n)
-        }
-        None => Err(FsError::InvalidArgument),
-    };
-    if result.is_ok() {
-        crate::kernel::syscall::io::wake_poll_waiters();
-    }
-    result
-}
-
-pub(crate) fn socket_write_impl(file: &SocketFile, buf: &[u8]) -> Result<usize, FsError> {
-    if file.is_shutdown_write() {
-        return Err(FsError::BrokenPipe);
-    }
-
-    let result = {
-        let mut sockets = SOCKET_SET.lock();
-        match file.handle.lock().as_ref() {
-            Some(SocketHandle::Tcp(h)) => {
-                let socket = sockets.get_mut::<tcp::Socket>(*h);
-                let result = socket.send_slice(buf).map_err(|_| FsError::WouldBlock);
-
-                if !buf.is_empty() {
-                    if let Ok(0) = result {
-                        if socket.may_send() {
-                            return Err(FsError::WouldBlock);
-                        } else {
-                            return Err(FsError::BrokenPipe);
-                        }
-                    }
-                }
-
-                result
-            }
-            Some(SocketHandle::Udp(h)) => {
-                let endpoint = match file.get_remote_endpoint() {
-                    Some(ep) => ep,
-                    None => return Err(FsError::NotConnected),
-                };
-                let socket = sockets.get_mut::<udp::Socket>(*h);
-                socket
-                    .send_slice(buf, endpoint)
-                    .map_err(|_| FsError::WouldBlock)?;
-                Ok(buf.len())
-            }
-            None => Err(FsError::InvalidArgument),
-        }
-    };
-    if result.is_ok() {
-        poll_network_interfaces_impl();
-        crate::kernel::syscall::io::wake_poll_waiters();
-    }
-    result
-}
-
-pub(crate) fn socket_recvfrom_impl(
-    file: &SocketFile,
-    buf: &mut [u8],
-) -> Result<(usize, Option<alloc::vec::Vec<u8>>), FsError> {
-    if file.is_shutdown_read() {
-        return Ok((0, None));
-    }
-
-    let mut sockets = SOCKET_SET.lock();
-    match file.handle.lock().as_ref() {
-        Some(SocketHandle::Tcp(h)) => {
-            let socket = sockets.get_mut::<tcp::Socket>(*h);
-            let state = socket.state();
-            if state == tcp::State::Closed {
-                return Ok((0, None));
-            }
-            if state == tcp::State::CloseWait && socket.recv_queue() == 0 {
-                return Ok((0, None));
-            }
-
-            let result = socket.recv_slice(buf).map_err(|_| FsError::WouldBlock);
-            let n = if let Ok(0) = result {
-                if state == tcp::State::CloseWait {
-                    0
-                } else if socket.may_recv() {
-                    return Err(FsError::WouldBlock);
-                } else {
-                    0
-                }
-            } else {
-                result?
-            };
-            let remote = socket.remote_endpoint().map(|ep| {
-                let mut addr_buf = alloc::vec![0u8; 16];
-                let _ = write_sockaddr_in_to_buf(&mut addr_buf, ep);
-                addr_buf
-            });
-            Ok((n, remote))
-        }
-        Some(SocketHandle::Udp(_)) => {
-            drop(sockets);
-            let Some(d) = file.udp_pop() else {
-                return Err(FsError::WouldBlock);
-            };
-            let n = core::cmp::min(buf.len(), d.len);
-            buf[..n].copy_from_slice(&d.data[..n]);
-            let mut addr_buf = alloc::vec![0u8; 16];
-            let _ = write_sockaddr_in_to_buf(&mut addr_buf, d.src);
-            Ok((n, Some(addr_buf)))
-        }
-        None => Err(FsError::InvalidArgument),
-    }
-}
-
-fn create_udp_socket_in_set(sockets: &mut SocketSet<'static>) -> Result<SmoltcpHandle, ()> {
-    // Allocate metadata buffers
-    let mut rx_meta_vec = alloc::vec::Vec::new();
-    rx_meta_vec.try_reserve(4).map_err(|_| ())?;
-    rx_meta_vec.resize(4, udp::PacketMetadata::EMPTY);
-
-    let mut tx_meta_vec = alloc::vec::Vec::new();
-    tx_meta_vec.try_reserve(4).map_err(|_| ())?;
-    tx_meta_vec.resize(4, udp::PacketMetadata::EMPTY);
-
-    // Allocate data buffers
-    let mut rx_data_vec = alloc::vec::Vec::new();
-    rx_data_vec.try_reserve(4096).map_err(|_| ())?;
-    rx_data_vec.resize(4096, 0);
-
-    let mut tx_data_vec = alloc::vec::Vec::new();
-    tx_data_vec.try_reserve(4096).map_err(|_| ())?;
-    tx_data_vec.resize(4096, 0);
-
-    let rx_buffer = udp::PacketBuffer::new(rx_meta_vec, rx_data_vec);
-    let tx_buffer = udp::PacketBuffer::new(tx_meta_vec, tx_data_vec);
-    let socket = udp::Socket::new(rx_buffer, tx_buffer);
-    Ok(sockets.add(socket))
-}
-
-pub(crate) fn create_tcp_socket_impl() -> Result<SocketHandle, ()> {
-    let mut rx_vec = alloc::vec::Vec::new();
-    rx_vec.try_reserve(4096).map_err(|_| ())?;
-    rx_vec.resize(4096, 0);
-
-    let mut tx_vec = alloc::vec::Vec::new();
-    tx_vec.try_reserve(4096).map_err(|_| ())?;
-    tx_vec.resize(4096, 0);
-
-    let rx_buffer = tcp::SocketBuffer::new(rx_vec);
-    let tx_buffer = tcp::SocketBuffer::new(tx_vec);
-    let socket = tcp::Socket::new(rx_buffer, tx_buffer);
-
-    let handle = SOCKET_SET.lock().add(socket);
-    Ok(SocketHandle::Tcp(handle))
-}
-
-pub(crate) fn create_udp_socket_impl() -> Result<SocketHandle, ()> {
-    let mut sockets = SOCKET_SET.lock();
-    let handle = create_udp_socket_in_set(&mut *sockets)?;
-    Ok(SocketHandle::Udp(handle))
 }
 
 const AF_INET: u16 = 2;
@@ -872,7 +361,7 @@ pub fn parse_sockaddr_in(addr: *const u8, addrlen: u32) -> Result<IpEndpoint, ()
 }
 
 /// Write sockaddr_in to buffer
-fn write_sockaddr_in_to_buf(buf: &mut [u8], endpoint: IpEndpoint) -> Result<(), ()> {
+pub(crate) fn write_sockaddr_in_to_buf(buf: &mut [u8], endpoint: IpEndpoint) -> Result<(), ()> {
     if buf.len() < SOCKADDR_IN_SIZE {
         return Err(());
     }
@@ -928,398 +417,36 @@ pub fn write_sockaddr_in(addr: *mut u8, addrlen: *mut u32, endpoint: IpEndpoint)
     Ok(())
 }
 
-/// Initialize network interface (should be called during network setup)
-pub(crate) fn init_network_impl(mut smoltcp_iface: crate::net::interface::SmoltcpInterface) {
-    let wrapper = NetIfaceWrapper {
-        device: SpinLock::new(smoltcp_iface.device_adapter_mut().clone()),
-        interface: SpinLock::new(smoltcp_iface.into_interface()),
-    };
-    *NET_IFACE.lock() = Some(wrapper);
-}
-
-/// Perform TCP connect with Context
-pub(crate) fn tcp_connect_impl(
-    handle: SmoltcpHandle,
-    remote: IpEndpoint,
-    local: IpEndpoint,
-) -> Result<(), ()> {
-    crate::pr_debug!("tcp_connect: start, handle={:?}", handle);
-
-    let iface_guard = NET_IFACE.lock();
-    crate::pr_debug!("tcp_connect: got NET_IFACE lock");
-
-    let wrapper = iface_guard.as_ref().ok_or(())?;
-
-    let result = wrapper.with_context(|context| {
-        crate::pr_debug!("tcp_connect: in with_context");
-        let mut sockets = SOCKET_SET.lock();
-        crate::pr_debug!("tcp_connect: got SOCKET_SET lock");
-        let socket = sockets.get_mut::<tcp::Socket>(handle);
-        crate::pr_debug!("tcp_connect: calling socket.connect");
-        let r = socket.connect(context, remote, local).map_err(|e| {
-            crate::pr_debug!("tcp_connect error: {:?}", e);
-            ()
-        });
-        crate::pr_debug!("tcp_connect: socket.connect returned {:?}", r);
-        r
-    });
-
-    // Poll immediately after connect to trigger SYN packet
-    if result.is_ok() {
-        crate::pr_debug!("tcp_connect: polling to send SYN");
-        wrapper.poll(&SOCKET_SET);
-    }
-
-    drop(iface_guard);
-    crate::pr_debug!("tcp_connect: done, result={:?}", result);
-    result
-}
-
-/// Poll network interfaces to process packets
-pub(crate) fn poll_network_interfaces_impl() {
-    if let Some(ref wrapper) = *NET_IFACE.lock() {
-        crate::pr_debug!("poll_network_interfaces: calling poll");
-        wrapper.poll(&SOCKET_SET);
-    }
-}
-
-/// Poll smoltcp + dispatch UDP datagrams to per-fd queues.
-///
-/// IMPORTANT: this may allocate (copies UDP payloads) and therefore must not be called from
-/// interrupt context.
-pub(crate) fn poll_network_and_dispatch_impl() {
-    poll_network_interfaces_impl();
-    if udp_dispatch_impl() {
-        crate::kernel::syscall::io::wake_poll_waiters();
-    }
-}
-
-/// Drain UDP datagrams from shared per-port sockets and deliver them to per-fd queues.
-///
-/// Returns whether any datagram was delivered.
-pub(crate) fn udp_dispatch_impl() -> bool {
-    let mut sockets = SOCKET_SET.lock();
-    udp_dispatch_drain_locked(&mut *sockets)
-}
-
-/// Attach an existing UDP fd to the shared per-port UDP socket, and register it as a logical socket
-/// for datagram dispatching.
-///
-/// Lock order: SOCKET_SET -> UDP_PORTS (must match NetIfaceWrapper::poll path).
-pub(crate) fn udp_attach_fd_to_port_impl(
-    tid: usize,
-    fd: usize,
-    file: &Arc<dyn crate::vfs::File>,
-    old_handle: SmoltcpHandle,
-    port: u16,
-    bind_addr: Option<IpAddress>,
-) -> Result<SmoltcpHandle, ()> {
-    // 1) Ensure shared per-port smoltcp socket exists and is bound.
-    let shared_handle = {
-        let mut sockets = SOCKET_SET.lock();
-        let mut ports = UDP_PORTS.lock();
-        if let Some(e) = ports.get(&port) {
-            e.handle
-        } else {
-            let h = create_udp_socket_in_set(&mut *sockets)?;
-            use smoltcp::wire::IpListenEndpoint;
-            let listen = IpListenEndpoint {
-                addr: bind_addr,
-                port,
-            };
-            if sockets.get_mut::<udp::Socket>(h).bind(listen).is_err() {
-                sockets.remove(h);
-                return Err(());
-            }
-            ports.insert(
-                port,
-                UdpPortEntry {
-                    handle: h,
-                    sockets: alloc::vec::Vec::new(),
-                },
-            );
-            h
-        }
-    };
-
-    // 2) Atomically switch fd mapping to the shared handle first (avoid stale-handle panics).
-    update_socket_handle(tid, fd, SocketHandle::Udp(shared_handle));
-    if let Some(sf) = file.as_any().downcast_ref::<SocketFile>() {
-        sf.set_handle(SocketHandle::Udp(shared_handle));
-    }
-
-    // 3) Register this logical socket for dispatch.
-    {
-        let mut ports = UDP_PORTS.lock();
-        if let Some(e) = ports.get_mut(&port) {
-            let already = e
-                .sockets
-                .iter()
-                .filter_map(|w| w.upgrade())
-                .any(|f| alloc::sync::Arc::ptr_eq(&f, file));
-            if !already {
-                e.sockets.push(alloc::sync::Arc::downgrade(file));
-            }
-        }
-    }
-
-    // 4) Finally remove the old per-fd smoltcp socket handle (it is unbound and should not receive).
-    if old_handle != shared_handle {
-        // Never remove a handle that is currently used as a shared per-port socket.
-        // (This can happen if user space calls bind/connect in an unexpected order.)
-        let mut sockets = SOCKET_SET.lock();
-        let ports = UDP_PORTS.lock();
-        let old_is_shared = ports.values().any(|e| e.handle == old_handle);
-        drop(ports);
-        if !old_is_shared {
-            sockets.remove(old_handle);
-        }
-    }
-
-    Ok(shared_handle)
-}
-
-/// Drain UDP datagrams from shared per-port sockets and deliver them to per-fd queues.
-///
-/// This function must be called with `SOCKET_SET` already locked.
-fn udp_dispatch_drain_locked(sockets: &mut SocketSet<'static>) -> bool {
-    let mut delivered_any = false;
-    let mut ports = UDP_PORTS.lock();
-    let mut to_remove: alloc::vec::Vec<(u16, SmoltcpHandle)> = alloc::vec::Vec::new();
-
-    for (port, entry) in ports.iter_mut() {
-        let socket = sockets.get_mut::<udp::Socket>(entry.handle);
-
-        while socket.can_recv() {
-            let (payload, meta) = match socket.recv() {
-                Ok(v) => v,
-                Err(_) => break,
-            };
-
-            let src = meta.endpoint;
-            let mut data = [0u8; UDP_DGRAM_MAX];
-            let copy_len = core::cmp::min(payload.len(), UDP_DGRAM_MAX);
-            data[..copy_len].copy_from_slice(&payload[..copy_len]);
-            let d = UdpDatagram {
-                src,
-                len: copy_len,
-                data,
-            };
-
-            // Prefer a connected socket that matches the remote endpoint. Otherwise, deliver to the
-            // first unconnected socket registered for this port.
-            let mut target: Option<alloc::sync::Arc<dyn crate::vfs::File>> = None;
-            let mut fallback: Option<alloc::sync::Arc<dyn crate::vfs::File>> = None;
-
-            entry.sockets.retain(|w| w.strong_count() > 0);
-
-            for w in entry.sockets.iter() {
-                let Some(f) = w.upgrade() else { continue };
-                let Some(sf) = f.as_any().downcast_ref::<SocketFile>() else {
-                    continue;
-                };
-
-                let Some(local_ep) = sf.get_local_endpoint() else {
-                    continue;
-                };
-                if local_ep.port != *port {
-                    continue;
-                }
-
-                match sf.get_remote_endpoint() {
-                    Some(remote) => {
-                        if remote.addr == src.addr && remote.port == src.port {
-                            target = Some(f.clone());
-                            break;
-                        }
-                    }
-                    None => {
-                        if fallback.is_none() {
-                            fallback = Some(f.clone());
-                        }
-                    }
-                }
-            }
-
-            let target = target.or(fallback);
-            if let Some(f) = target {
-                if let Some(sf) = f.as_any().downcast_ref::<SocketFile>() {
-                    if sf.udp_push(d) {
-                        delivered_any = true;
-                    }
-                }
-            }
-        }
-
-        // Prune dead logical sockets; if none remain, remove the shared smoltcp socket.
-        entry.sockets.retain(|w| w.strong_count() > 0);
-        if entry.sockets.is_empty() {
-            to_remove.push((*port, entry.handle));
-        }
-    }
-
-    for (port, handle) in to_remove {
-        ports.remove(&port);
-        sockets.remove(handle);
-    }
-
-    delivered_any
-}
-
-/// Poll until loopback queue is empty
-pub(crate) fn poll_until_empty_impl() {
-    if let Some(ref wrapper) = *NET_IFACE.lock() {
-        // Always poll at least once to process socket state changes
-        wrapper.poll(&SOCKET_SET);
-
-        // Then drain loopback queue, but do it in bounded steps.
-        //
-        // Draining until empty can take unbounded time when user programs (e.g. iperf3 UDP)
-        // generate packets faster than the stack can consume them, causing apparent "hangs".
-        const MAX_DRAIN_POLLS: usize = 256;
-        for _ in 0..MAX_DRAIN_POLLS {
-            if wrapper.loopback_queue_len() == 0 {
-                break;
-            }
-            wrapper.poll(&SOCKET_SET);
-        }
-    }
-}
-
 pub fn create_tcp_socket() -> Result<SocketHandle, ()> {
-    crate::net::stack::network_stack()
-        .lock()
-        .create_tcp_socket()
+    crate::net::stack::network_stack().create_tcp_socket()
 }
 
 pub fn create_udp_socket() -> Result<SocketHandle, ()> {
-    crate::net::stack::network_stack()
-        .lock()
-        .create_udp_socket()
+    crate::net::stack::network_stack().create_udp_socket()
 }
 
 /// Initialize network interface through the stack facade.
 pub fn init_network(smoltcp_iface: crate::net::interface::SmoltcpInterface) {
-    crate::net::stack::network_stack()
-        .lock()
-        .init_network(smoltcp_iface);
+    crate::net::stack::network_stack().init_network(smoltcp_iface);
 }
 
 pub fn tcp_connect(handle: SmoltcpHandle, remote: IpEndpoint, local: IpEndpoint) -> Result<(), ()> {
-    crate::net::stack::network_stack()
-        .lock()
-        .tcp_connect(handle, remote, local)
-}
-
-pub fn tcp_connection_state_impl(
-    handle: SmoltcpHandle,
-) -> Option<crate::net::stack::TcpConnectionState> {
-    let sockets = SOCKET_SET.lock();
-    let socket = sockets.get::<tcp::Socket>(handle);
-    Some(match socket.state() {
-        tcp::State::Established => crate::net::stack::TcpConnectionState::Established,
-        tcp::State::Closed => crate::net::stack::TcpConnectionState::Closed,
-        _ => crate::net::stack::TcpConnectionState::Other,
-    })
-}
-
-pub(crate) fn tcp_listen_impl(
-    handle: SmoltcpHandle,
-    endpoint: smoltcp::wire::IpListenEndpoint,
-) -> Result<(), ()> {
-    let mut sockets = SOCKET_SET.lock();
-    sockets
-        .get_mut::<tcp::Socket>(handle)
-        .listen(endpoint)
-        .map_err(|_| ())
-}
-
-pub(crate) fn tcp_listener_state_endpoint_impl(
-    handle: SmoltcpHandle,
-) -> Option<(
-    crate::net::stack::TcpListenState,
-    smoltcp::wire::IpListenEndpoint,
-)> {
-    let sockets = SOCKET_SET.lock();
-    let socket = sockets.get::<tcp::Socket>(handle);
-    let state = match socket.state() {
-        tcp::State::Listen => crate::net::stack::TcpListenState::Listen,
-        tcp::State::Established => crate::net::stack::TcpListenState::Established,
-        tcp::State::CloseWait => crate::net::stack::TcpListenState::CloseWait,
-        _ => crate::net::stack::TcpListenState::Other,
-    };
-    Some((state, socket.listen_endpoint()))
-}
-
-pub(crate) fn remove_tcp_socket_impl(handle: SmoltcpHandle) {
-    SOCKET_SET.lock().remove(handle);
-}
-
-pub(crate) fn tcp_accept_endpoints_impl(
-    handle: SmoltcpHandle,
-) -> Option<(IpEndpoint, Option<IpEndpoint>)> {
-    let sockets = SOCKET_SET.lock();
-    let socket = sockets.get::<tcp::Socket>(handle);
-    let remote = socket.remote_endpoint()?;
-    Some((remote, socket.local_endpoint()))
-}
-
-pub(crate) fn tcp_debug_state_impl(
-    handle: SmoltcpHandle,
-) -> Option<(crate::net::stack::TcpConnectionState, bool)> {
-    let sockets = SOCKET_SET.lock();
-    let socket = sockets.get::<tcp::Socket>(handle);
-    let state = match socket.state() {
-        tcp::State::Established => crate::net::stack::TcpConnectionState::Established,
-        tcp::State::Closed => crate::net::stack::TcpConnectionState::Closed,
-        _ => crate::net::stack::TcpConnectionState::Other,
-    };
-    Some((state, socket.is_open()))
-}
-
-pub(crate) fn tcp_close_impl(handle: SmoltcpHandle) {
-    SOCKET_SET.lock().get_mut::<tcp::Socket>(handle).close();
-}
-
-pub(crate) fn socket_local_endpoint_impl(handle: SocketHandle) -> Option<IpEndpoint> {
-    let sockets = SOCKET_SET.lock();
-    match handle {
-        SocketHandle::Tcp(h) => sockets.get::<tcp::Socket>(h).local_endpoint(),
-        SocketHandle::Udp(h) => {
-            let listen_ep = sockets.get::<udp::Socket>(h).endpoint();
-            Some(IpEndpoint::new(
-                listen_ep
-                    .addr
-                    .unwrap_or(IpAddress::Ipv4(Ipv4Address::UNSPECIFIED)),
-                listen_ep.port,
-            ))
-        }
-    }
-}
-
-pub(crate) fn socket_remote_endpoint_impl(handle: SocketHandle) -> Option<IpEndpoint> {
-    let sockets = SOCKET_SET.lock();
-    match handle {
-        SocketHandle::Tcp(h) => sockets.get::<tcp::Socket>(h).remote_endpoint(),
-        SocketHandle::Udp(_) => None,
-    }
+    crate::net::stack::network_stack().tcp_connect(handle, remote, local)
 }
 
 /// Poll network interfaces to process packets.
 pub fn poll_network_interfaces() {
-    crate::net::stack::network_stack().lock().poll();
+    crate::net::stack::network_stack().poll();
 }
 
 /// Poll smoltcp + dispatch UDP datagrams to per-fd queues.
 pub fn poll_network_and_dispatch() {
-    crate::net::stack::network_stack()
-        .lock()
-        .poll_and_dispatch();
+    crate::net::stack::network_stack().poll_and_dispatch();
 }
 
 /// Drain UDP datagrams from shared per-port sockets and deliver them to per-fd queues.
 pub fn udp_dispatch() -> bool {
-    crate::net::stack::network_stack().lock().udp_dispatch()
+    crate::net::stack::network_stack().udp_dispatch()
 }
 
 pub fn udp_attach_fd_to_port(
@@ -1331,11 +458,10 @@ pub fn udp_attach_fd_to_port(
     bind_addr: Option<IpAddress>,
 ) -> Result<SmoltcpHandle, ()> {
     crate::net::stack::network_stack()
-        .lock()
         .udp_attach_fd_to_port(tid, fd, file, old_handle, port, bind_addr)
 }
 
 /// Poll until loopback queue is empty.
 pub fn poll_until_empty() {
-    crate::net::stack::network_stack().lock().poll_until_empty();
+    crate::net::stack::network_stack().poll_until_empty();
 }
